Switch nox job default back to Jammy
When I was trying to get Zuul to test against MariaDB by default I
switched the base test node to bookworm so that py311 and MariaDB 10.11
would be used together. The reason for this is it seemed like maybe
MariaDB 10.6 on Jammy was buggy and causing our connection errors. Since
then we've discovered that 10.11 and 10.6 share the same underlying bug
which we worked around via not flushing privileges.

Switch these tests back to Jammy as this will simplify the process of
dogfooding zuul-launcher (one less image to worry about).

Change-Id: I1ed4252617e5570522838e87dca4d8caefb8849e