Annotate "reporting start|enqueue|dequeue" log messages with event UUID
This annotates the log messages for start, enqueue and dequeue reporting
with the Zuul event UUID based on the QueueItem's original event.

This UUID might differ from the UUID of the event that triggered the
reporting. This is e.g. the case when a change is dequeued or promoted
via the Zuul admin API.

However, as those events are handled asynchronously in the scheduler,
this is the case for most log messages that are "invoked" by handling
management events.

Having these log messages annotated with the QueueItem's event UUID
should aid debugging, especially when using the logs to view the whole
lifecycle of a QueueItem while it's processed by Zuul.

Change-Id: I0f93ddf7bdd4a0c27ed28dfc279148a17bda27c9