Use eatmydata with zookeeper in tests
This uses eatmydata with zookeeper when we run our tests.  This
should allow us to avoid using a tmpfs (which uses memory) for
ZK, which generally shouldn't need to read back any data in our
unit tests, so we don't really care how long background writes
take.  This should save 3/4 of a GB of ram.

Change-Id: I64ef1e9028d836c91d1c7c59909ac1f5a54db2a5