Always clear wake event of TPE event connectors
During shutdown of the event connectors we can still have a significant
number of pending event futures. So far, when the connector was stopped
we set the stop flag and wake event to notify the dispatcher.

The problem we encountered was, that the dispatcher thread started
busy-looping and hogging the CPU as the wake event was never cleared
when the stop flag was set. This was caused by the wake event's
`wait(0.1)` always returning immediately.

To fix that we will always clear the wake event at the start of the
dispatcher loop.

Change-Id: I68832c386214921fb61e1227ac18fd93994d2ef6