Make test_zuul_query_ltime more deterministic
We have seen some test runs where test_zuul_query_ltime submits
3 events at once, yet the events are processed in two clusters
rather than one.  The intent of the test is for all three events
to be procesesd simultaneously to exercise the ChangeNetwork that
allows us to reduce gerrit queries.

To ensure that all three events are performing queries simultaneously,
this change adds some threading primitives to ensure the exact
sequencing of the three threads involved.

Change-Id: I3c8584a9b1a5dad1d7584b271b452e1628d8a660