Use contextlib.nullcontext across the board
We had a vendored version of contextlib.nullcontext for python older
than 3.7. We don't support python that old anymore so we drop the
vendored context manager entirely and pull it from stdlib. While I was
cleaning this up I noticed the merger also had a nullcontext
implementation that I have also replaced with the stdlib variant.

Change-Id: I29e8af4682e4dbd3268a2e407565c236e6f4a385