Omit succssful jobs on very long gerrit reports
If the gerrit report is very long (like, hundreds of jobs) it may
exceed the 16k message length limit and will be truncated.  If only
a small number failed, then the failed jobs may be in the truncated
section and not visible to users.  To avoid that situation, if we
find that we are going to truncate the report, then reformat it with
successful jobs omitted.  That will ensure that users see the failed
jobs.

If we end up with a very long report with only successful jobs, this
does mean that we will omit the entire job listing.  That's probably
okay since there should still be a buildset link in the header.

Change-Id: I3d538ead162b2132bd3ee9ca9dfcfc8a78ac7999