Improve job variant descriptions
If a job variant has a Zuul Regex as a branch matcher, we will use
the JSON serialized version of the regex for the tab header.  That
is not pretty and awkward to read, so reformat it to be a more
textual description: it will now either be the regex or the regex
followed by the string "(negated)" if that is the case.

Change-Id: Ie47cb39acb76e9bc2451cab28f8fec24ae5e9fb8