Store resources of pod type nodes in host inventory
Ansibles built-in `ansible_*` facts for compute resources are not cgroup
aware, ie., they report the resources of the host machine, rather than
the limits set by k8s for the container we're running on.

While this information might still be useful, we also want to know in
later analyzing steps what the real resource limits of a build were.
Therefore, include the resources for a pod node in the host inventory
together with other k8s related facts as stored by nodepool.

Change-Id: Ib16b3f4bc21966b86b4081317d575fe571cf7c22