Allow boto3 session to configure itself
So far we've required either a AWS shared credentials file or access key
to configure the AWS connection.

However, there are other valid ways to configure a boto3 session (e.g.
via AWS web identity token file) which we also want to enable.

To support this, we will use the AWS access key ID and secret access key
from the zuul.conf - if provided - but rely on boto3 to configure itself
for all other methods.

The AWS connection will use the Security Token Service to validate that
the session works as expected. The `get_caller_identity` API which is
used for that doesn't require any permissions [0].

[0] https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/sts/client/get_caller_identity.html

Change-Id: I79d29f33f5eb26b97fb17660edc0d44a84c40ed9