Always include project name in change key
Gerrit Code Review has deprecated the use of legacy ids many
releases ago [1]; the ways to identify a change are:

- <project>~<changeNumber>
- <project>~<branch>~<Change-Id>

The other form of IDs, <Change-Id> or <ChangeNumber> alone,
do not guarantee to uniquely identify a single change; also,
when there is no ambiguity and they are enough to identify
a change, they do perform quite badly due to the additional
lookup through the secondary index, which could be under
pressure.

The recommended way to identify a change is to use
<project>~<changeNumber>, therefore the ChangeKey needs
to store also the project that is always available in the
event object.

[1] https://gerrit-documentation.storage.googleapis.com/Documentation/3.11.2/rest-api-changes.html#change-id

Change-Id: I353b48ef18818eec9ddee4b1abde0e84208ee96f