Avoid 500 on /project with no project configs
If a project is added to a tenant and there are no project configs
for that tenant, then there will be no project metadata on the layout.
In that case, return an abbreviated result with as much information
as we have (almost none).  But it will be enough to indicate that
the project does exist in the tenant (ie, not a 404), but with an
empty list of configs and an empty metadata dict.

Change-Id: Id29b152e48d1c75d1f66fd960dc4c9bc35a6cf7d