Add niz cleanup
This adds leaked resource cleanup, using rendezvous hashing to
resolve which launcher should clean each endpoint.  There is
no secondary locking since a collision on cleanup is not critical.

This adds more testing than nodepool had for leaked fip and port
cleanup in openstack.  Options to configure cleanup behavior
are added.

Note that the options are set in the providers, but they really
affect endpoints.  This is an unavoidable mismatch, so we
use the most generous values.  This means that any openstack
provider on an endpoint can enable fip/port cleanup for
that endpoint.

The AWS tests used different regions which caused issues with
resources being partitioned.  Standardize on us-east-1 since
that was what most used.

AWS cleanup requires a bucket name to clean up leaked image
upload objects.  This can be configured on a provider, and is
an unavoidable mismatch for endpoint cleanups.  To resolve this,
we look at the bucket names for all providers for an endpoint
when we perform resource cleanup.

The shutdown sequence for moto is adjusted so that it happens
at the end of the normal test shutdown sequence.

The system_id is added to the provider and endpoint objects so
that they can construct an appropriate metadata dictionary
at any point.

Change-Id: I373df424e7db3d5ac22cdcd2d29a243ed960201d