Slow processing of failed nodescan connections
This change ports the missing parts of
Ic3f5b223fafd805d2357171e7822e0e33bd104bb to Zuul.

The nodescan processing loop iterates over the state machine for
every pending request and attempts to advance it every time the loop
runs.  The loop runs either every second or every time a socket is
ready for io.

If a connection is failing quickly, we attempt to avoid a busy-wait
by removing its socket and resetting the state machine to START so
that it will only open a new connection after the 1 second timeout.

But if there are a lot of connections (or, perhaps even just more
than one that is failing) then we will go through the loop quickly
anyway, and we may effectively busy-wait.

To avoid this, we will process all connections with a ready socket
any time they are ready.  But we will process the remaining
connections only once every second.  An exception is made for the
case where we can move a request from the pending queue to the active
queue so we don't unecessarily delay the start of processing.

Additionally, if a connection fails during the first key exchange,
reset to START since it's still safe to do so, and we will get the
benefit of the extra delay.

Any connection failures after the first key exchange will still
re-connect immediately.

Change-Id: I49fcadd94c3bac36a980135e5048039c573ae801