Wake launcher when node is (un-)locked
When an executor returns the nodeset it will first set the provider node
state to 'used' and unlock the node afterwards. If the launcher tries to
lock the node for cleanup before it is unlocked by the executor, the
node won't be processed until there is some other event that wakes up
the launcher.

To prevent this edge-case we will wake up the launcher if the lock state
of a node changes.

Change-Id: I640e3ade60d85b641e8cd32191e45588412200bb