Demote change list refresh error to warning
We try an opportunistic refresh of the change list while processing
the tenant trigger queue.  We do this without locking, which means
that we may try loading it while another actor is writing, and
we would receive an error.  We previously anticipated encountering
a JSON decode error and recording that as a warning rather than an
error with a traceback, but it's also possible for the individual
znode shards to be messing which would raise an exception.

Since any exception is not likely actionable by the user and we
proceed regardless, and also because we later perform a refresh
of the change list in a locked context where any persistent structural
error would become visible (and we would still log an exception in
that case), let's treat all errors here as warnings.

In addition to improving the UX when reading the logs, this will
reduce some false positives in test races.

Change-Id: Iceb9ca940e68736dd07e10febfbaa276d5fffc21