Fix issue with parse context w/o scheduler
As of I373df424e7db3d5ac22cdcd2d29a243ed960201d the configloader
needs the Zuul system ID for getting a provider.

However, the scheduler attribute of the parse context is only valid on
schedulers, but e.g. not zuul-web. Since we need the system ID when
adding providers to the layout, we need to pass the `ZuulSystem` as a
separate argument.

Change-Id: I3585edbaad934531f88e6ba51ecaaa6cf7f16836