Gerrit: retry fetching refs and HEAD
Make the Gerrit driver more robust against transient network errors
by retrying requests to get the list of refs (which we use for
branch listings) and the default branch.  Failures here can be
particularly problematic because they can leave a tenant in an old
configuration state, and require manual intervention to fix.

This change also makes a minor adjustment to the existing retry
handling for gerrit queries so that the loop iteration matches
other retry loops in the driver, and we also log the retries as
warnings so that operators can better monitor for problems.

Change-Id: I5fdad2532764d4af8dacb25dd3515d0c7e47e49d