Fix exception handling in Gerrit event connector
The original intent for connection event handling is that if there
is an error with the event, it should be discarded.  A recent
change to the Gerrit driver to use a threadpool executor to
pre-process events in parallel inadvertently altered the error
handling for that driver so that if certain exceptions were
encountered, the event would be retained in the incoming queue
indefinitely.

To return to the old behavior, the exception handling in the
gerrit event processor is updated so that all exceptions result
in the event handler completing sucessfully and the original
connection event is removed from the queue (regardless of whether
it produces any trigger events).

The new form looks similar to the corresponding code in the
github driver, which has used this structure for a longer time.

Change-Id: I0465a5c00e49300b13b75db626e0dc80a89cf4a1