Add newrev to timer-triggered items
If jobs in a periodic pipeline run longer than the timer interval
for the pipeline trigger, Zuul will not enqueue a second queue
item for the same project-branch.  This can be counter-intuitive
for users, where a pipeline that runs every 24 hours may turn into
a pipeline that runs every 48 hours if a job happens to take 25
hours.

To address this, we will now store the branch head sha as the
newrev of the ref associated with queue items.  This will cause
periodic pipelines to behave a bit more like post pipelines, in
that multiple items will be allowed for the same project-branch.
However, if the branch sha has not changed, we will not enqueue
further items (since they will continue to look identical to
Zuul).

Change-Id: I41891ac3011fb95c4c891c7554e1ea6bec94b56f