Adjust niz private_ipv4 behavior to match Nodepool
Nodepool would set the private_ipv4 value to either private_ipv4
or public_ipv4 depending on which existed.  We should maintain
backwards-compatability to aid in the transition, but we should
also clean up the interface.  This change does both.

A new zuul.node dictionary is added to be a future replacement
for the nodepool dict.  In that, we use the new clean API.

Change-Id: I5e554dc4cb231f973c5c5cadeb8e78f87ed72780