Silently dequeue changes when handling promote events
When handling promote events for Github changes, Zuul also updates the
check run on Github side as "cancelled" when dequeue reporting is
configured for the pipeline. However, we don't do another start
reporting as start reporting is only done once (when the initial
buildset is enqueued).

As a result, the check run in Github will stay in cancelled until the
new buildset is completed and directly switch over to the final
"success" or "failure" result. In the meantime there are no updates on
this check run, giving the impression to the user that Zuul just
cancelled the buildset but didn't restart anything.

In order to fix this, silently dequeue changes when handling promote
events. According to our test suite this will still create a new check
run on Github side (most probably because we end up with a new buildset)
and leave the old check run in running state. We should validate this
behaviour with a real Github instance first to make sure it's not
related to our fakegithub implementation.

Change-Id: I4698d53375e4c3287af75deea74a8270d3662a62