Fix AtributeError when canceling merge requests
When trying to cancel a merge request, Zuul raises the following
exception:

    Traceback (most recent call last):
      File "/opt/zuul/lib/python3.11/site-packages/zuul/configloader.py", line 2500, in _cacheTenantYAML
        self.merger.cancel(cancel_job)
      File "/opt/zuul/lib/python3.11/site-packages/zuul/merger/client.py", line 264, in cancel
        request = self.merger_api.get(job.request_path)
                  ^^^^^^^^^^^^^^^^^^^
    AttributeError: 'MergerApi' object has no attribute 'get'

The MergerApi.get() method was renamed to MergerApi.getRequest() in
Ie3d62dadd9990c847df8302d4da801a990394d2c, but apparently this call in
the cancel procedure was forgotten.

Since getRequest is using the job UUID rather than the request_path, we
first have to extract the UUID from the path.

This enhances the TestingMergerApi to allow explicitly failing merge
jobs for a given job type. To make this work properly we also have to
use the TestingMergerApi on the executor server that is used within the
Zuul tests.

Change-Id: Ib2f81959202adc08aff5eb905f3d73b10cd2891c