Make '--tenant' optional in create-auth-token command
The zuul.admin claim is not always wanted when creating token using
zuul-admin create_auth_token command using custom claims.

Fix: when --tenant is not specified it would not add zuul.admin claim
to the token.

Mean while, print the custom claims when --print-meta-info is
specified, that they are also important meta info.

Change-Id: I21b9c83bf3655586efc61d95c40c4adc63be8bff