WIP Add greater control to tag branch matchers
Since 04ac8287b6 git tags have matched the branches that they are
contained within. This commit notes that when a single tag is contained
within multiple branches it will match all of those branches possibly
producing unexpected behavior when unwanted/unexpected job variants
apply.

This change is a bit of a request for comment on an approach to give
Zuul users in this situation an out. Specifically, we allow for
refs/tags branch matchers to short circuit and bail out of containing
branches processing. As the included test illustrates this allows you to
define explicit variants that match specific tags and they won't simply
be included by default beacuse they are contained within the branch the
tag commit lives in.

Prior to this change any job variants defined within containing branches
would apply to tags contained within those branches. The reason for that
is we didn't short circuit on negative matches and would fall through
the containing branches handling.

As with 04ac8287b6 I suspect there may still be unexpected behavior and
corner cases with this update. But I figured getting something posted so
that we can discuss a concrete change would be helpful.

Change-Id: I1e26ddd90a0e5e89564b6f63c5632c63c0c36c72