Fix issue with config override validators
It looks like re-creating the schema on every invocation of the override
list (and probably also the override value) validator can lead to issues
during multi-threaded config loading.

This can result in random config errors for attributes using the
override validators, e.g. during parsing of a project template:

    no valid value found @ data['check']['jobs'][105]['foobar']['dependencies'][0]

What seems to fix the issue is to create the schema once when the
builder function is called and not recreate/compile it with every
invocation of the validator.

Change-Id: I29c2dfa4e2109a7c4da7174a418f895c49b00720