Fix issue with config override validators
Re-creating the schema on every invocation of the override list (and
probably also the override value) validator can lead to issues during
multi-threaded config loading.

This can result in random config errors for attributes using the
override validators, e.g. during parsing of a project template:

    no valid value found @ data['check']['jobs'][105]['foobar']['dependencies'][0]

This is because compiling a vs.Any validator into a schema is not
thread-safe for the validator object (because it stores the compiled
form on itself).

The override_list method was called with a vs.Any validator as input
in some instances (e.g., `override_list(vs.Any(ZUUL_REGEX, str))`.

When the validator closure function within override_list compiled its
internal vs.Any validator, it also compiled the vs.Any validator
supplied as input.  This means that any time the validator closure
method was called, it might temporarily empty the list of compiled
validators on the supplied input validator.

The solution is to avoid compiling validators within validation
methods.

Therefore we will now create the schema once when the builder function
is called and will not recreate/compile it with every invocation of
the validator.

Change-Id: I29c2dfa4e2109a7c4da7174a418f895c49b00720