Delete nodeset requests for canceled jobs
The niz code path for deleting nodeset requests for canceled jobs
was unimplemented.  Add an implementation that mirrors the nodepool
version.

Also, remove an unecessary check of the type of the node request
id which was left over from the old implementation of node request
deduplication (since superceded by the circular dep refactor).

Change-Id: Iebb16ae3cf441b66de63bc3468e5723c7580de80