Add RendezvousElection
This adds a new zk helper class, the RendezvousElection.

This behaves like a kazoo Election, but it uses rendezvous hashing
to determine who the winner should be.  It chooses the winner among
all of the "running" components of the specified type by comparing
a hash of the components' hostnames plus a string representing the
object of the election (the lock path in this case).

A lock is still used to ensure that only one component will win
the election at a time if components are out of sync, or a component
comes on or offline.

Users of this election can periodically check whether it is still
valid and gracefully exit their run methods to yield to different
component in the case that adding or removing a component creates
a new winner.  This is important when a system is brought online:
in that case we may see one component added at a time.  The first
component would win all the elections, but as more are added, the
system will begin to distribute load among them.

The initial consumer of this is expected to be the connection event
processors which are run on the scheduler.  For this case, we want
any "running" scheduler to be used since all schedulers are
interchangeable.  If other components, such as executor, use this
in the future, they may need to filter the set by other criteria
(such as zone), so a customizable filter option is supplied.

Change-Id: I3b0bd24f88fe5313a5f699d898d9fefacad855ba