Handle s3 signed urls in launcher
The AWS s3 service is a likely location for image upload artifacts.
It supports access control and sharing via pre-signed urls (like swift),
but in that case does not implement the HEAD method.  To work around
that, if we encounter a 403, we will retry with a zero-byte range GET
request.  This request, like HEAD, will return no data, but will supply
the total number of bytes in the content-range header.

Change-Id: I66f1a2e7e79d44b7e3215d5e2f6b994569b746cc