Allow for overriding python_path in the testsuite
Zuul has a subset of test cases that actually run Ansible against fake
tests nodes that are actually just localhost using the local Ansible
connection. When jobs run Ansible against these fake nodes it runs
Ansible setup using the nullwrapper which auto discovers the python path
for the execution environment. Depending on what platform you run this
may be the system python or the virtualenv python running the test
suite. Next the job runs the freeze playbook which uses the bubblwrap
wrapper in a more limited environment that does not contain the
virtualenv python in its bind mount list.

Ansible caches the python path discovered by the setup playbook and
reuses it for subsequent playbooks. If the running platform is not known
to Ansible then it discovers the first python in its path which is
likely the virtualenv python. It then later tries to run the freeze
playbook with this python which fails due to that python not being bind
mounted into the bwrap environment. This occurs on platforms Ansible
doesn't know about like OpenSUSE Tumbleweed, but does not happen on more
common platforms like Debuntu. On Debuntu Ansible looks up a table for
known python path installations and prefers those which end up being the
system python which is available inside and outside of the bubblewrap
context.

To make it possible for users of platforms like OpenSUSE Tumbleweed
(like myself) to run the full Zuul testsuite we update the test tooling
to look for ZUUL_ANSIBLE_PYTHON env var values which will be used to
override the test node python_path. We only do this in the test cases
that actually run Ansible.

Change-Id: I89bcd9fff0785dfc27e2381b15e13533223088bd