Fix config errors on always-dynamic branches
The change to cache configuration objects had an error which could
cause errors on always-dynamic branches to appear when they should not.

If a branch of a repo is included normally in one tenant, and included
in another tenant as an always-dynamic branch, and that branch has a
configuration error, then the configuration error would be included
in any dynamic layouts generated.

This is because the _loadDynamicProjectData method examines all branches
in every project it is called for.  Previously when it would add objects
from the cache, it would pull from the cache on the tenant project
config, which only included objects relevant to that tenant.  Since it
now pulls from the global config object cache, we need to ensure that
we only use those objects on non-dynamic branches.

In other words, _loadDynamicProjectData will now
* get all the branches for the project, including dynamic branches
* check to see if we have files entries for each branch
  * if we do, parse and include objects from the files entries
  * if we do not, then use objects from the cache, but only
    if the branch is not dynamic.

We can't simply start with a list of branches that excludes dynamic
branches, because we still need to check those for the presense of
files entries from the merger.  But once that is complete, we can
disregard dynamic branches.

Change-Id: I8851f6b3d4207fce5d3dcdbf354941801b827e78