Fix 'is_locked' flag for items in ZKObject cache
This change fixes a problem with the LockableZKObjectCache cache and the
`is_locked` indicator.

The problem was that the pre-cache hook did not take into account
multiple lock contender nodes and was setting the `is_locked` flag to
false when any of the contenders was deleted.

To fix this the items in the cache will keep a set of lock contenders
and consider the node locked as long as contenders exist.

Change-Id: I4341498b09b34ada39fe19e6ceab2f66187b7361