Add AWS direct image import
Two of the AWS methods for uploading images (image import and
snapshot import) operate on objects in S3.  We can save a
download/upload round trip on the launcher by directly importing
the s3 object that was previously uploaded by the zuul image
build job.

This will only work if the object is in S3, and we will need to
parse the bucket and object name out of the URL.  To accomodate
this, we will only enable this feature if the returned URL is
an "s3://" url, since that is easily parseable.

Since this only works under certain conditions, we need to be
able to ask the driver if it can perform this before we decide
whether to download the image.  To handle this, we encapsulate
all of the pre-flight checks in new provider methods.  These
methods will determine if the provider can perform an import, and
if so, will return an "ImageUploadJob" object which effectively
acts like a closure that can later be run in an executor.

In this way, we can iterate over all of the uploads we need to
perform, get jobs for all of the ones that can be directly imported,
then, if any uploads remain which can not be directly imported,
we can download the image and collect jobs for regular file-based
uploads.

The existing uploadImage base provider method is replaced with
two new methods: getImageImportJob and getImageUploadJob.  In a
provider, implementing the first is optional, and the second is
required iff the provider supports images.

Change-Id: Iae00b6b8d6595681bd54ca84ce1aef6fa5ad8a4e