Use non-validated images for image validation jobs
Implement the missing functionality to request and use a non-validated
upload for the validation job.

In case of a image validation trigger we will pass the image names and
upload UUID via the nodeset request to the launcher. If those properties
are set the launcher will provision a node with the requested image.

This is only relevant if the requested label is actually using the image
that needs to be validated.

TODO:
* cleanup
* extend test
* clarify if we want to fully "initialize" the AWS mocks for the
  test_launcher test cases
* clarify what's the use-case for passing a list of image names with the
  image-validation trigger as AFAIKS there is a 1:1 rel. between an
  upload and an image

Change-Id: Ic09720c7bf2a0709aba90eac1b7383096285b580