Support downloading images from s3
This allows users to upload images to non-public s3 buckets and
use Zuul's authenticated connection to download the images.

If the zuul-launcher is unable to download a supplied url using
python requests, it will allow each provider in the tenant to
attempt to download the image using its own internal methods.
The only method implemented currently is the s3 download method.

Change-Id: I54fbf5ba23743ed0ddf42e37ac1c0536df69b2c7