Allow enqueing duplicate changes if they are for different images
When we enqueue items for missing images, we try very hard to
enqueue one item with all of the images listed as part of the
event triggering the queue item (therefore, all the jobs for
the missing images will run).  Likewise, a periodic pipeline would
enqueue a single queue item that builds all the images for
that project.

However, if we manually enqueue an image build, then any other
image build events for that project will be ignored as long as
that item is in the queue.  This is because the "change" for the
item is, like a periodic pipeline, the branch head for that project.
Since we only allow one live copy of a change in a pipeline at a
time, subsequent events are ignored.

The same behavior also causes problems for image validation events,
which are emitted one at a time.

To correct this and provide a more intuitive behavior, in the case
that a queue item has image names attached, we will check to see
if the existing item represents a superset of the images requested
in the new item, and if it is not, we will allow the new item to
be enqueued as well.

Change-Id: I2745978bce62b1ea247277316cf50cb28385fe17