Add stdout/err to streaming log if console disabled
If the user sets zuul_console_disabled to true, we know that we
will not get live logs for command, etc, tasks.  In that case,
get the output from the result dict and add it to the logs after
the task completes.

This can make it easier for users to perform a phased upgrade
to windows log streaming.  It can also improve the experience
for both windows and unix users who can't run zuul-console for
any reason.

Change-Id: I5d3e885377b1e158ad8a932da1023c77d15a7dac