Report cherrypy stats/metrics at least every 60 seconds
Previously the cherrypy stats/metrics were only reported when processing
a new request. This means that as long as we are busy we would collect
accurate info but as the system becomes idle we're left with stale
information. This is problematic if you want to use these stats/metrics
for automated scaling of the zuul web service. In particular you won't
get accurate info indicating when you can scale down in all situations.
By ensuring we report up to date info every 60 seconds we should now
report when we are ilde which can be used to scale down decisions.

Another thing this addresses is the idea of a ttl on the gauge data. If
a server is scaled and goes down it will no longer report stats. When we
did not consistently report stats on a regular interval we couldn't
accurately expire that data in say prometheus or graphite. Now we can.

Change-Id: I2d228261f56906d8f4e854d525824205e7f37d71