Improve zuul_console_disabled testing
This adds a function test that verifies:
* That we really do avoid creating a connection to the log streamer
  when zuul_console_disabled is set.

  This is accomplished by ensuring that we don't get a
  "Starting to log..." message in the ansible-playbook command output.
* That we do include the task output in the console log text file
  when zuul_console_disabled is set (we copy it from the task result
  dictionary).

The existing tests around this functionality did not have effective
assertions for the first item, and did not effectively exercise it
because they set zuul_console_disabled as a task variable.  We expect
it to be set as a host variable (since it is a characteristic of the
host), so the test is updated to use set_fact instead.

(It would be ideal to also handle it as a task variable, but it's not
clear if there is a reasonable way to do that in the callback plugin.)

Change-Id: I780e265b4d86ace7dd06d577a44ed5a61fe2a740