wip: Only copy compatible defaults from provider config
In order to allow mixing cloud and zuul images in the same provider
config that also includes image defaults we need to make sure to only
use defaults from the provider that are applicable for a particular
image type.

Otherwise having Zuul image settings on the provider level will cause
config errors for clouds images.

Change-Id: I3b3e3ca9423da30450764f6fbbe0ad28c7f6a56d