Ensure job projects are only considered once
When a project is already part of the required-projects for a particular
job we don't want to include it again if it is in addition needed for
include-vars.

Adding a project twice could lead to a wrong workspace checkout in case
the override-checkout and change branch are different. Due to evaluation
order only the include-vars were taken from the override-checkout
branch.

To fix this we will make sure to not include the vars project again if
it is already in the list of projects.

Change-Id: I1e51f41db21634c36d14a2b06184b628623ff4a5