WIP: Handle out-of-memory host failures
When an Ansible process on a build node (mostly K8s Pod nodes) gets
OOMKilled, we treat this currently as a regular FAILURE result, without
further notice to the user about the reason (despite the 137 exit code
in the Ansible output). When this happens in a pre-playbook, we even
retry those jobs. But given a statically configured memory limit for
those Pod build nodes, they will likely just keep failing until they
reach RETRY_LIMIT.

This is not only a waste of resources, but users might not even know
that the error lies in excessive memory consumption and they need to
adjust their label config before attempting another run.

We can capture these cases by parsing the Ansible output for child
processes (ie. kubectl) exiting with the 137 exit code, which is
commonly used to indicate that a process was killed by the oom-killer
[1].

With this change, if we encounter such a log message after the Ansible
process was returning with exit code 2 (indicating a failing host), we
skip any job retries and return a FAILURE build result together with an
"out-of-memory" error detail.

[1] The exit code 137 generally indicates that a process has received a
SIGKILL. So we might run into cases where we report "out-of-memory" but
it wasn't in fact the OOMKiller that sent the SIGKILL. We cannot detect
the source of the SIGKILL. A closer approximation would be to limit this
log line detection to nodes of connection_type "kubectl".

Change-Id: Ifa81431b6581de416c68d29dba2e107bf6f0ae38