Detect and report out-of-memory host failures
When an Ansible process on a build node (mostly K8s Pod nodes) gets
OOMKilled, we treat this currently as a FAILURE result, without further
notice to the user about the reason (despite the 137 exit code in the
Ansible output).

We can capture these cases by parsing the Ansible output for child
processes (ie. kubectl) exiting with the 137 exit code, which is
commonly used to indicate that a process was killed by the oom-killer
[1].

With this change, if we encounter such a log message after the Ansible
process was returning with exit code 2 (indicating a failing host), we
treat this as an "unreachable" build result together with an
"out-of-memory" error detail.

[1] The exit code 137 generally indicates that a process has received a
SIGKILL. So we might run into cases where we report "out-of-memory" but
it wasn't in fact the OOMKiller that sent the SIGKILL. We cannot detect
the source of the SIGKILL. A closer approximation would be to limit this
log line detection to nodes of connection_type "kubectl".

Change-Id: Ifa81431b6581de416c68d29dba2e107bf6f0ae38